move logging.py inside utils
