+++ movement/__init__.py
@@ -1,6 +1,6 @@
 from importlib.metadata import PackageNotFoundError, version
 
-from movement.logging import configure_logging
+from movement.utils.logging import configure_logging
 from movement.move_accessor import MovementDataset
 
 try:
