+++ movement/io/load_poses.py
@@ -12,7 +12,7 @@ from sleap_io.io.slp import read_labels
 from sleap_io.model.labels import Labels
 
 from movement import MovementDataset
-from movement.logging import log_error, log_warning
+from movement.utils.logging import log_error, log_warning
 from movement.validators.datasets import ValidPosesDataset
 from movement.validators.files import ValidDeepLabCutCSV, ValidFile, ValidHDF5
