+++ movement/io/save_poses.py
@@ -9,7 +9,7 @@ import numpy as np
 import pandas as pd
 import xarray as xr
 
-from movement.logging import log_error
+from movement.utils.logging import log_error
 from movement.validators.files import ValidFile
 
 logger = logging.getLogger(__name__)
